Implement builder pattern for bptree serialize process
