+++ src/blob/index/bptree/core.rs
@@ -14,7 +14,7 @@
 ///    where pointer is offset in file of underlying node with searched key
 use super::prelude::*;
 
-const BLOCK_SIZE: usize = 4096;
+pub(super) const BLOCK_SIZE: usize = 4096;
 
 #[derive(Debug, Clone)]
 pub(crate) struct BPTreeFileIndex {
@@ -48,11 +48,7 @@ impl FileIndexTrait for BPTreeFileIndex {
         recreate_index_file: bool,
     ) -> Result<Self> {
         let res = Self::serialize(headers, filter)?;
-        if res.is_none() {
-            error!("Indices are empty!");
-            return Err(anyhow!("empty in-memory indices".to_string()));
-        }
-        let (mut header, metadata, buf) = res.expect("None case is checked");
+        let (mut header, metadata, buf) = res;
         clean_file(path, recreate_index_file)?;
         let file = File::create(path, ioring)
             .await
@@ -277,158 +273,11 @@ impl BPTreeFileIndex {
     fn serialize(
         headers_btree: &InMemoryIndex,
         filter: &Bloom,
-    ) -> Result<Option<(IndexHeader, TreeMeta, Vec<u8>)>> {
-        debug!("blob index bptree serialize headers");
-        if let Some(record_header) = headers_btree.values().next().and_then(|v| v.first()) {
-            let record_header_size = record_header.serialized_size().try_into()?;
-            let filter_buf = filter.to_raw()?;
-            let headers_len = headers_btree.iter().fold(0, |acc, (_k, v)| acc + v.len());
-            let header = IndexHeader::new(record_header_size, headers_len, filter_buf.len());
-            let header_buf = serialize(&header)?;
-            let hs = header_buf.len() as usize;
-            let fsize = header.filter_buf_size;
-            let msize: usize = TreeMeta::serialized_size_default()? as usize;
-            let headers_start_offset = (hs + fsize) as u64;
-            let headers_size = headers_len * record_header_size;
-            let leaves_offset = headers_start_offset + (headers_size + msize) as u64;
-            // leaf contains pair: (key, offset in file for first record's header with this key)
-            let leaf_size = record_header.key().len() + std::mem::size_of::<u64>();
-            let leaves_buf = Self::serialize_leaves(
-                headers_btree,
-                headers_start_offset,
-                leaf_size,
-                record_header_size,
-            );
-            let keys = headers_btree.keys().collect();
-            let tree_offset = leaves_offset + leaves_buf.len() as u64;
-            let (root_offset, tree_buf) =
-                Self::serialize_bptree(keys, leaves_offset, leaf_size as u64, tree_offset)?;
-            let metadata = TreeMeta::new(root_offset, leaves_offset, tree_offset);
-            let meta_buf = serialize(&metadata)?;
-
-            let data_size = hs + fsize + headers_size + msize + leaves_buf.len() + tree_buf.len();
-            let mut buf = Vec::with_capacity(data_size);
-            serialize_into(&mut buf, &header)?;
-            buf.extend_from_slice(&filter_buf);
-            Self::append_headers(headers_btree, &mut buf)?;
-            buf.extend_from_slice(&meta_buf);
-            buf.extend_from_slice(&leaves_buf);
-            buf.extend_from_slice(&tree_buf);
-            let hash = get_hash(&buf);
-            let header =
-                IndexHeader::with_hash(record_header_size, headers_len, filter_buf.len(), hash);
-            serialize_into(buf.as_mut_slice(), &header)?;
-            Ok(Some((header, metadata, buf)))
-        } else {
-            Ok(None)
-        }
-    }
-
-    fn append_headers(headers_btree: &InMemoryIndex, buf: &mut Vec<u8>) -> Result<()> {
-        headers_btree
-            .iter()
-            .flat_map(|r| r.1)
-            .map(|h| serialize(&h))
-            .try_fold(buf, |buf, h_buf| -> Result<_> {
-                buf.extend_from_slice(&h_buf?);
-                Ok(buf)
-            })?;
-        Ok(())
-    }
-
-    fn serialize_bptree(
-        keys: Vec<&Vec<u8>>,
-        leaves_offset: u64,
-        leaf_size: u64,
-        tree_offset: u64,
-    ) -> Result<(u64, Vec<u8>)> {
-        let max_amount = Self::max_leaf_node_capacity(keys[0].len());
-        let nodes_amount = (keys.len() - 1) / max_amount + 1;
-        let elems_in_node = (keys.len() / nodes_amount) as u64;
-        trace!(
-            "last node has {} less keys",
-            elems_in_node - elems_in_node % nodes_amount as u64
-        );
-
-        let leaf_nodes_compressed = keys
-            .chunks(elems_in_node as usize)
-            .enumerate()
-            .map(|(i, keys)| {
-                let cur_offset = leaves_offset + elems_in_node * leaf_size * i as u64;
-                (keys[0].clone(), cur_offset)
-            })
-            .collect::<Vec<(Vec<u8>, u64)>>();
-
-        let mut buf = Vec::new();
-        let root_offset = Self::build_tree(leaf_nodes_compressed, tree_offset, &mut buf)?;
-        Ok((root_offset, buf))
-    }
-
-    pub(super) fn build_tree(
-        nodes_arr: Vec<(Vec<u8>, u64)>,
-        tree_offset: u64,
-        buf: &mut Vec<u8>,
-    ) -> Result<u64> {
-        if nodes_arr.len() == 1 {
-            return Ok(Self::prep_root(nodes_arr[0].1, tree_offset, buf));
-        }
-        let max_amount = Self::max_nonleaf_node_capacity(nodes_arr[0].0.len());
-        let nodes_amount = (nodes_arr.len() - 1) / max_amount + 1;
-        let elems_in_node = nodes_arr.len() / nodes_amount;
-        let new_nodes = nodes_arr
-            .chunks(elems_in_node)
-            .map(|keys| {
-                let offset = tree_offset + buf.len() as u64;
-                let min_key = keys[0].0.clone();
-                let offsets = keys.iter().map(|(_, offset)| *offset).collect();
-                let keys = keys.iter().skip(1).map(|(key, _)| key.clone()).collect();
-                let node = Node::new(keys, offsets);
-                buf.extend_from_slice(&node.serialize().expect("failed to serialize node"));
-                (min_key, offset)
-            })
-            .collect();
-        Self::build_tree(new_nodes, tree_offset, buf)
-    }
-
-    fn prep_root(offset: u64, tree_offset: u64, buf: &mut Vec<u8>) -> u64 {
-        let root_node_size = if offset < tree_offset {
-            (tree_offset - offset) as usize
-        } else {
-            buf.len() - (offset - tree_offset) as usize
-        };
-        let appendix_size = BLOCK_SIZE - root_node_size;
-        buf.resize(buf.len() + appendix_size, 0);
-        offset
-    }
-
-    fn max_leaf_node_capacity(key_size: usize) -> usize {
-        let offset_size = std::mem::size_of::<u64>();
-        BLOCK_SIZE / (key_size + offset_size) - 1
-    }
-
-    fn max_nonleaf_node_capacity(key_size: usize) -> usize {
-        let offset_size = std::mem::size_of::<u64>();
-        let meta_size = std::mem::size_of::<NodeMeta>();
-        (BLOCK_SIZE - meta_size - offset_size) / (key_size + offset_size) + 1
-    }
-
-    fn serialize_leaves(
-        headers_btree: &InMemoryIndex,
-        headers_start_offset: u64,
-        leaf_size: usize,
-        record_header_size: usize,
-    ) -> Vec<u8> {
-        let mut leaves_buf = Vec::with_capacity(leaf_size * headers_btree.len());
-        headers_btree.iter().fold(
-            (headers_start_offset, &mut leaves_buf),
-            |(offset, leaves_buf), (leaf, hds)| {
-                leaves_buf.extend_from_slice(&leaf);
-                let offsetb = serialize(&offset).expect("serialize u64");
-                leaves_buf.extend_from_slice(&offsetb);
-                let res = (offset + (hds.len() * record_header_size) as u64, leaves_buf);
-                res
-            },
-        );
-        leaves_buf
+    ) -> Result<(IndexHeader, TreeMeta, Vec<u8>)> {
+        Serializer::new(headers_btree)
+            .header_stage(filter)?
+            .leaves_stage()?
+            .tree_stage()?
+            .build()
     }
 }
