+++ src/blob/index/bptree/mod.rs
@@ -3,14 +3,16 @@ use super::prelude::*;
 mod core;
 mod meta;
 mod node;
+mod serializer;
 #[cfg(test)]
 mod tests;
 
 pub(crate) use self::core::BPTreeFileIndex;
 
 mod prelude {
+    pub(super) use super::core::BLOCK_SIZE;
+    pub(super) use super::serializer::Serializer;
     pub(super) use super::*;
-    pub(super) use meta::{TreeMeta, NodeMeta};
+    pub(super) use meta::{NodeMeta, TreeMeta};
     pub(super) use node::Node;
 }
-
